Add files via upload
Created graph in R of average height by year. There appears to be differences in data before-and-after 2011. This may be due to a large die-off in 2011, but also earlier data may have mostly recorded data by 5 classifications or had limited data for previous years. Using that data in the future will be required for better analysis.